review: Move tags formatters to the namespace TagsFormatter and rename SupportedTagsFormatter.
