+++ src/JustEat.StatsD/Buffered/StatsDUtf8Formatter.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Text;
-using JustEat.StatsD.Buffered.Tags;
+using JustEat.StatsD.TagsFormatters;
 
 namespace JustEat.StatsD.Buffered
 {
