+++ src/JustEat.StatsD/TagsFormatters/NoOpTagsFormatter.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace JustEat.StatsD.Buffered.Tags
+namespace JustEat.StatsD.TagsFormatters
 {
     internal sealed class NoOpTagsFormatter : IStatsDTagsFormatter
     {
