+++ tests/JustEat.StatsD.Tests/Utf8FormatterTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Text;
 using JustEat.StatsD.Buffered;
-using JustEat.StatsD.Buffered.Tags;
+using JustEat.StatsD.TagsFormatters;
 using Shouldly;
 using Xunit;
